pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["PolyakovDmitry546 <polyakov1803@gmail.com>"]
readme = "README.md"
packages = [{include = "django_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
pillow = "^9.5.0"
django-mptt = "^0.14.0"
django-ckeditor = "^6.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
